# enable but inact memory profiling