--- a/components/playground/instance/process.go
+++ b/components/playground/instance/process.go
@@ -2,6 +2,7 @@ package instance
 
 import (
 	"context"
+	"fmt"
 	"io"
 	"os"
 	"os/exec"
@@ -84,8 +85,8 @@ func (p *process) Cmd() *exec.Cmd {
 	return p.cmd
 }
 
-// NewComponentProcess create a Process instance.
-func NewComponentProcess(ctx context.Context, dir, binPath, component string, version utils.Version, arg ...string) (Process, error) {
+// NewComponentProcessWithEnvs create a Process instance with given environment variables.
+func NewComponentProcessWithEnvs(ctx context.Context, dir, binPath, component string, version utils.Version, envs map[string]string, arg ...string) (Process, error) {
 	if dir == "" {
 		panic("dir must be set")
 	}
@@ -95,20 +96,31 @@ func NewComponentProcess(ctx context.Context, dir, binPath, component string, ve
 
 	env := environment.GlobalEnv()
 	params := &tiupexec.PrepareCommandParams{
-		Ctx:         ctx,
-		Component:   component,
-		Version:     version,
-		BinPath:     binPath,
-		InstanceDir: dir,
-		WD:          dir,
-		Args:        arg,
-		SysProcAttr: SysProcAttr,
-		Env:         env,
+		Ctx:          ctx,
+		Component:    component,
+		Version:      version,
+		BinPath:      binPath,
+		InstanceDir:  dir,
+		WD:           dir,
+		Args:         arg,
+		EnvVariables: make([]string, 0),
+		SysProcAttr:  SysProcAttr,
+		Env:          env,
 	}
+	for k, v := range envs {
+		pair := fmt.Sprintf("%s=%s", k, v)
+		params.EnvVariables = append(params.EnvVariables, pair)
+	}
+
 	cmd, err := tiupexec.PrepareCommand(params)
 	if err != nil {
 		return nil, err
 	}
 
 	return &process{cmd: cmd}, nil
+}
+
+// NewComponentProcess create a Process instance.
+func NewComponentProcess(ctx context.Context, dir, binPath, component string, version utils.Version, arg ...string) (Process, error) {
+	return NewComponentProcessWithEnvs(ctx, dir, binPath, component, version, nil, arg...)
 }