--- a/embed/templates/scripts/run_tikv.sh.tpl
+++ b/embed/templates/scripts/run_tikv.sh.tpl
@@ -20,6 +20,8 @@ echo $stat
   {{- end}}
 {{- end}}
 
+export MALLOC_CONF="prof:true,prof_active:true,prof.active:false"
+
 {{- if .NumaNode}}
 exec numactl --cpunodebind={{.NumaNode}} --membind={{.NumaNode}} bin/tikv-server \
 {{- else}}