--- a/pkg/exec/run.go
+++ b/pkg/exec/run.go
@@ -127,17 +127,18 @@ func base62Tag() string {
 
 // PrepareCommandParams for PrepareCommand.
 type PrepareCommandParams struct {
-	Ctx         context.Context
-	Component   string
-	Version     utils.Version
-	BinPath     string
-	Tag         string
-	InstanceDir string
-	WD          string
-	Args        []string
-	SysProcAttr *syscall.SysProcAttr
-	Env         *environment.Environment
-	CheckUpdate bool
+	Ctx          context.Context
+	Component    string
+	Version      utils.Version
+	BinPath      string
+	Tag          string
+	InstanceDir  string
+	WD           string
+	Args         []string
+	EnvVariables []string
+	SysProcAttr  *syscall.SysProcAttr
+	Env          *environment.Environment
+	CheckUpdate  bool
 }
 
 // PrepareCommand will download necessary component and returns a *exec.Cmd
@@ -222,6 +223,7 @@ func PrepareCommand(p *PrepareCommandParams) (*exec.Cmd, error) {
 		fmt.Sprintf("%s=%s", localdata.EnvTag, p.Tag),
 	}
 	envs = append(envs, os.Environ()...)
+	envs = append(envs, p.EnvVariables...)
 
 	// init the command
 	c := exec.CommandContext(p.Ctx, binPath, p.Args...)