Signed-off-by: qupeng <qupeng@pingcap.com>

### What problem does this PR solve? <!--add issue link with summary if exists-->
Close #1270 and https://github.com/tikv/tikv/issues/9538.

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Manual test (add detailed scripts or steps below)
   1. Start a TiKV cluster with the environment variable set.
   2. wget "http://172.16.4.78:50013/debug/pprof/heap?seconds=1" to dump a memory profiling

Code changes

Side effects

Related changes

 - Need to cherry-pick to the release branch
 - Need to update the documentation

Release notes:

```release-note
NONE
```