@lucklove 2021-04-01
To fix https://github.com/tikv/tikv/issues/9538, maybe you want to change `tiup-playground` instead of `tiup-cluster`?