@hicqu 2021-04-01
> To fix [tikv/tikv#9538](https://github.com/tikv/tikv/issues/9538), maybe you want to change `tiup-playground` instead of `tiup-cluster`?

@lucklove all TiKV clusters have the problem, regardless they are deployed with `playground` or `cluster`.

This PR is only for `cluster`, do you have any idea for setting environment variables for `playground`?