@lucklove 2021-04-01
> > To fix [tikv/tikv#9538](https://github.com/tikv/tikv/issues/9538), maybe you want to change `tiup-playground` instead of `tiup-cluster`?
> 
> @lucklove all TiKV clusters have the problem, regardless they are deployed with `playground` or `cluster`.
> 
> This PR is only for `cluster`, do you have any idea for setting environment variables for `playground`?

TiUP set envs for each component [here](https://github.com/pingcap/tiup/blob/master/pkg/exec/run.go#L224). 

1. You can add a `Envs []string` filed for [PrepareCommandParams](https://github.com/pingcap/tiup/blob/master/pkg/exec/run.go#L129) 
2. And add `envs = append(envs, p.Envs)` on L224.
3. Then add envs arg for [NewComponentProcess](https://github.com/pingcap/tiup/blob/master/components/playground/instance/process.go#L88)
4. Add your env when call [NewComponentProcess in TiKVInstance](https://github.com/pingcap/tiup/blob/master/components/playground/instance/tikv.go#L74)