@hicqu 2021-04-01
I have benchmarked it, here is the result:
```
without any environment variables:
$ cargo bench
running 7 tests
test bench_malloc_128k ... bench:          39 ns/iter (+/- 7)
test bench_malloc_128m ... bench:       4,134 ns/iter (+/- 412)
test bench_malloc_1g   ... bench:       8,478 ns/iter (+/- 2,356)
test bench_malloc_1k   ... bench:          48 ns/iter (+/- 5)
test bench_malloc_1m   ... bench:          37 ns/iter (+/- 5)
test bench_malloc_8k   ... bench:          30 ns/iter (+/- 4)
test bench_malloc_8m   ... bench:          35 ns/iter (+/- 3)
```
```
with `export MALLOC_CONF="prof:true,prof_active:true,prof.active:false"`
$ cargo bench
running 7 tests
test bench_malloc_128k ... bench:          37 ns/iter (+/- 16)
test bench_malloc_128m ... bench:       4,136 ns/iter (+/- 605)
test bench_malloc_1g   ... bench:       8,426 ns/iter (+/- 1,134)
test bench_malloc_1k   ... bench:          44 ns/iter (+/- 6)
test bench_malloc_1m   ... bench:          35 ns/iter (+/- 3)
test bench_malloc_8k   ... bench:          32 ns/iter (+/- 3)
test bench_malloc_8m   ... bench:          37 ns/iter (+/- 5)
```

There aren't obvious dirrerences.

The reason is memory profiling is compiled for both 2 cases, and it's not actived in both 2 cases.